requirements.txt
certifi==2022.6.15
charset-normalizer==2.1.0
cloudinary==1.29.0
confusable-homoglyphs==3.2.0
dj-database-url==0.5.0
dj3-cloudinary-storage==0.0.6
Django==2.2.10
django-appconf==1.0.4
django-ckeditor==6.0.0
django-crispy-forms==1.9.2
django-heroku==0.3.1
django-imagekit==4.0.2
django-js-asset==1.2.2
django-material==1.4.3
django-registration==3.1.1
django-registration-defaults==0.4.4
gunicorn==20.1.0
idna==3.3
image==1.5.32
pilkit==2.0
Pillow==8.0.0
psycopg2==2.8.6
python-decouple==3.4
pytz==2020.1
requests==2.28.1
six==1.15.0
sqlparse==0.3.1
urllib3==1.26.10
whitenoise==5.3.0
